refactor: change type ID Activity Group to int64

+++ models/domain/activity_group.go
@@ -3,7 +3,7 @@ package domain
 import "time"
 
 type ActivityGroup struct {
-	ID        int        `gorm:"primary_key"`
+	ID        int64      `gorm:"primary_key"`
 	Email     string     `gorm:"type:varchar(191);not null;unique"`
 	Title     string     `gorm:"type:varchar(191);not null"`
 	CreatedAt *time.Time `gorm:"autoCreateTime"`
